chore: move `cargo_near::commands::build_command::build::run` fn to `cargo_near_lib::build`
