+++ cargo-near-build/Cargo.toml
@@ -19,3 +19,4 @@ serde_json = "1.0"
 near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
 zstd = "0.13"
 schemars = "0.8"
+rustc_version = "0.4"
