+++ cargo-near-build/src/types/near/build.rs
@@ -1,9 +1,13 @@
 use std::marker::PhantomData;
 
 use camino::Utf8PathBuf;
+use colored::{ColoredString, Colorize};
 use sha2::{Digest, Sha256};
 
-use crate::{types::color_preference::ColorPreference, ArtifactType, WASM};
+use crate::{
+    cargo_native::{ArtifactType, WASM},
+    types::color_preference::ColorPreference,
+};
 
 pub mod version_mismatch;
 
@@ -109,5 +113,37 @@ impl Opts {
             .into_iter()
             .map(|el| el.to_string())
             .collect::<Vec<_>>()
+    }
+}
+
+// TODO: make non-pub
+#[derive(Default)]
+pub struct ArtifactMessages<'a> {
+    messages: Vec<(&'a str, ColoredString)>,
+}
+
+impl<'a> ArtifactMessages<'a> {
+    pub fn push_binary(&mut self, artifact: &CompilationArtifact) -> eyre::Result<()> {
+        self.messages
+            .push(("Binary", artifact.path.to_string().bright_yellow().bold()));
+        let checksum = artifact.compute_hash()?;
+        self.messages.push((
+            "SHA-256 checksum hex ",
+            checksum.to_hex_string().green().dimmed(),
+        ));
+        self.messages.push((
+            "SHA-256 checksum bs58",
+            checksum.to_base58_string().green().dimmed(),
+        ));
+        Ok(())
+    }
+    pub fn push_free(&mut self, msg: (&'a str, ColoredString)) {
+        self.messages.push(msg);
+    }
+    pub fn pretty_print(self) {
+        let max_width = self.messages.iter().map(|(h, _)| h.len()).max().unwrap();
+        for (header, message) in self.messages {
+            eprintln!("     - {:>width$}: {}", header, message, width = max_width);
+        }
     }
 }
